Pull request #3865: [BLUETOOTH-16945] APP: Fix for CERT-C
Merge in CONNINT/bluetooth_public from bugfix/BLUETOOTH-16945-unresolved-msg-findings-middleware_wireless_bluetooth to develop/main

Squashed commit of the following:

commit bcfa49cfcd971e14d2f3f62ff2df6efcd4184f1c
Author: Alexandru Vrabete <alexandru.vrabete@nxp.com>
Date:   Mon Mar 2 15:27:09 2026 +0200

    [BLUETOOTH-16945] APP: Fix for CERT-C

+++ application/common/ble_port_fsci/core0/nvm/app_nvm_core0.c
@@ -642,7 +642,7 @@ static bleResult_t App_NvmRead
                                 FLib_MemCpy(pRamData, *ppNvmData, mSize);
                                 pRamData = (void *)((uint8_t *)pRamData + mSize);
                                 *pDataSetBitmask |= nvmId_BondingDataDescriptorBit_c;
-                                *pDescriptorBitmask |= (uint8_t)(1U << descIdx);
+                                *pDescriptorBitmask |= (1UL << descIdx);
                             }
 
                             tempDescBitmask &= tempDescBitmask - 1U;
